# openapi.yaml
openapi: 3.1.0
info:
  title: 校园二手物品交易系统 API
  version: 1.0.0
  description: |
    本文件是 MVP 全部 HTTP 接口的唯一契约（Task-ID: API-01），依据总纲 v2.6 §3/§4/§7/§8 生成。
    实现任务与契约不一致时，先修订本文件与总纲，再改实现。

    通用约定：
    - 基础路径 `/api/v1` 由 `servers` 提供，paths 中不重复该前缀；
    - 除 `POST /auth/wechat-login`、`POST /auth/refresh` 与标注为公开的接口外，所有操作要求 Bearer Token；
    - 成功响应统一包裹 `ApiResponse`，失败响应统一包裹 `ApiError`，HTTP 状态与错误码的映射见总纲 §7.2；
    - ID（含嵌套 ID 与 bizId）一律为十进制字符串；金额与平均分为两位小数字符串；时间为 UTC ISO 8601 毫秒；
    - 返回 204 的操作没有响应体；
    - 每个操作的 `x-owner-task` 声明实现它的任务 ID。
servers:
  - url: /api/v1
    description: 应用基础路径
...
security:
  - bearerAuth: []
paths:
...
  /orders/{id}:
    get:
      tags: [order]
      operationId: getOrderDetail
      x-owner-task: BE-08
      summary: 获取订单详情
      description: 仅买家或卖家可见，否则 `ORDER_OPERATION_FORBIDDEN`；管理员不是订单的天然参与方，同样需要是买家或卖家才可见（管理端的全量订单视图走管理端接口）。使用下单时的不可变快照字段，返回事件时间线与 `allowedActions`。前端按 `allowedActions` 展示按钮，后端每次仍重新校验。
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: 订单详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
...
  /admin/orders:
    get:
      tags: [admin]
      operationId: listAdminOrders
      x-owner-task: QA-01
      summary: 分页查询全站订单（管理员）
      description: 管理员专属只读视图，与管理员账号自己的买入/卖出列表分离。支持状态、买家、卖家及订单号/商品标题关键字筛选。
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: buyerId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DecimalId'
        - name: sellerId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DecimalId'
        - name: keyword
          in: query
          required: false
          description: 订单号或商品标题关键字，1～64 字符
          schema:
            type: string
            minLength: 1
            maxLength: 64
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 全站订单分页
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageResponse'
                          - type: object
                            required: [items]
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/AdminOrderSummary'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
  /admin/orders/{id}:
    get:
      tags: [admin]
      operationId: getAdminOrderDetail
      x-owner-task: QA-01
      summary: 获取全站订单只读详情（管理员）
      description: 管理员专属读取任意订单；返回快照与时间线，但 `allowedActions` 固定为空，不能借管理身份执行买卖双方动作。
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: 管理员订单详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        登录或刷新接口返回的 15 分钟 Access Token，放在 `Authorization: Bearer <accessToken>`。
        Token 无效或过期返回 `AUTH_UNAUTHORIZED`；前端不得对登录/刷新接口本身启用自动刷新逻辑。

  # ===========================================================================
  # 复用参数
  # ===========================================================================
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: |
        调用方生成的请求标识。只接受 1～64 个 ASCII 字母、数字、下划线或连字符；
        缺失或非法时服务端重新生成。响应头与响应体中的 `requestId` 一致。
      schema:
        $ref: '#/components/schemas/RequestId'
    Page:
      name: page
      in: query
      required: false
      description: 页码，从 0 开始
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: 每页条数
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    OrderId:
      name: id
      in: path
      required: true
      description: 订单 ID（十进制字符串）
      schema:
        $ref: '#/components/schemas/DecimalId'
...
  responses:
    Error400:
      description: 请求参数、查询参数或请求体不合法
      x-error-codes:
        - VALIDATION_ERROR
        - FILE_INVALID_TYPE
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Error401:
      description: 未认证或凭证无效
      x-error-codes:
        - AUTH_INVALID_CODE
        - AUTH_UNAUTHORIZED
        - AUTH_REFRESH_INVALID
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Error403:
      description: 已认证但无权限，或账号状态不允许该操作
      x-error-codes:
        - AUTH_FORBIDDEN
        - USER_DISABLED
        - USER_CERTIFICATION_REQUIRED
        - FILE_NOT_OWNED
        - ORDER_OPERATION_FORBIDDEN
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Error404:
      description: 资源不存在，或对当前用户不可见
      x-error-codes:
        - RESOURCE_NOT_FOUND
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Error500:
      description: 未分类的服务端异常，响应不得暴露内部细节
      x-error-codes:
        - INTERNAL_ERROR
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  # ===========================================================================
  # 数据模型
  # ===========================================================================
  schemas:
...
    RequestId:
      type: string
      pattern: '^[A-Za-z0-9_-]{1,64}$'
      minLength: 1
      maxLength: 64
      description: 请求标识，1～64 个 ASCII 字母、数字、下划线或连字符
      example: 01JXXXXXXXXXXXX
    DecimalId:
      type: string
      pattern: '^(0|[1-9][0-9]*)$'
      description: |
        十进制字符串形式的 ID。请求、路径与响应一致使用字符串，
        避免微信 JavaScript 丢失 BIGINT 精度。分页计数等非 ID 整数仍是 JSON number。
      example: '1024'
    UtcInstant:
      type: string
      format: date-time
      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z$'
      description: UTC 时间，ISO 8601，固定毫秒三位并以 `Z` 结尾
      example: '2026-09-10T04:50:00.000Z'
    MoneyAmount:
      type: string
      pattern: '^(0|[1-9][0-9]*)\.[0-9]{2}$'
      description: 两位小数的十进制字符串金额。服务端转为 BigDecimal，不使用浮点
      example: '10.00'
...
    ErrorCode:
      type: string
      description: §7.2 固定错误码全集
      enum:
        - VALIDATION_ERROR
        - AUTH_INVALID_CODE
        - AUTH_UNAUTHORIZED
        - AUTH_REFRESH_INVALID
        - AUTH_FORBIDDEN
        - USER_DISABLED
        - USER_CERTIFICATION_REQUIRED
        - RESOURCE_NOT_FOUND
        - CERTIFICATION_PENDING_EXISTS
        - CERTIFICATION_ALREADY_REVIEWED
        - CATEGORY_IN_USE
        - FILE_INVALID_TYPE
        - FILE_TOO_LARGE
        - FILE_NOT_OWNED
        - ITEM_NOT_EDITABLE
        - ITEM_NOT_AVAILABLE
        - ITEM_SELF_PURCHASE
        - ITEM_CONCURRENTLY_RESERVED
        - ITEM_SELF_OPERATION
        - ORDER_ILLEGAL_STATUS_TRANSITION
        - ORDER_OPERATION_FORBIDDEN
        - ORDER_DUPLICATE_REQUEST
        - REVIEW_NOT_ALLOWED
        - REVIEW_ALREADY_EXISTS
        - INTERNAL_ERROR
    ApiResponse:
      type: object
      description: 所有带响应体的成功响应都使用该信封；HTTP 204 无响应体。
      required: [code, message, requestId, timestamp]
      properties:
        code:
          type: string
          const: OK
          description: 成功固定为 `OK`
        message:
          type: string
          description: 成功固定为 `success`
        data:
          description: 业务数据。具体结构由各响应声明；失败响应固定为 null。
        requestId:
          $ref: '#/components/schemas/RequestId'
        timestamp:
          $ref: '#/components/schemas/UtcInstant'
    ApiError:
      type: object
      description: 错误响应体，HTTP 状态码与 `code` 按 §7.2 一一对应。
      required: [code, message, data, requestId, timestamp]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: 可读说明。不得包含堆栈、SQL、文件路径、Token 或任何敏感字段。
        data:
          type: 'null'
          description: 错误响应固定为 null
        requestId:
          $ref: '#/components/schemas/RequestId'
        timestamp:
          $ref: '#/components/schemas/UtcInstant'
    PageResponse:
      type: object
      description: 分页元数据。列表元素类型由各响应通过 allOf 补充。
      required: [page, size, totalElements, totalPages, hasNext]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean

    # -------------------------------------------------------------------------
    # §4 枚举（15 个）
    # -------------------------------------------------------------------------
...
    TradeMode:
      type: string
      enum: [OFFLINE, SIMULATED_PAYMENT]
      description: |
        TradeMode（§4）。MVP 只接受 `OFFLINE`；`SIMULATED_PAYMENT` 保留在领域枚举中，
        不作为已支持流程。
    OrderStatus:
      type: string
      enum: [PENDING_CONFIRMATION, CONFIRMED, PENDING_RECEIPT, COMPLETED, CANCELLED, REJECTED]
      description: OrderStatus（§4）
    OrderAction:
      type: string
      enum: [CREATE, CONFIRM, REJECT, CANCEL, DELIVER, RECEIVE]
      description: OrderAction（§4）
...
    UserSummary:
      type: object
      description: 出现在其他资源中的用户精简摘要，永不包含 openid
      required: [id, nickname]
      properties:
        id:
          $ref: '#/components/schemas/DecimalId'
        nickname:
          type: string
        avatarUrl:
          type: [string, 'null']
          description: 头像访问地址；为空表示未设置
...
    OrderItemSnapshot:
      type: object
      description: 下单时固化的商品快照，商品后续修改不影响订单展示
      required: [itemId, title, price]
      properties:
        itemId:
          $ref: '#/components/schemas/DecimalId'
        title:
          type: string
        imageUrl:
          type: [string, 'null']
        price:
          $ref: '#/components/schemas/MoneyAmount'
    OrderEvent:
      type: object
      description: 只追加的订单事件
      required: [id, action, toStatus, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/DecimalId'
        action:
          $ref: '#/components/schemas/OrderAction'
        fromStatus:
          anyOf:
            - $ref: '#/components/schemas/OrderStatus'
            - type: 'null'
          description: 创建事件没有前置状态
        toStatus:
          $ref: '#/components/schemas/OrderStatus'
        operator:
          $ref: '#/components/schemas/UserSummary'
        remark:
          type: [string, 'null']
        createdAt:
          $ref: '#/components/schemas/UtcInstant'
...
    OrderDetail:
      type: object
      required: [id, orderNo, status, tradeMode, amount, item, buyer, seller, allowedActions, events, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/DecimalId'
        orderNo:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        tradeMode:
          $ref: '#/components/schemas/TradeMode'
        amount:
          $ref: '#/components/schemas/MoneyAmount'
        cancelReason:
          type: [string, 'null']
        item:
          $ref: '#/components/schemas/OrderItemSnapshot'
        buyer:
          $ref: '#/components/schemas/UserSummary'
        seller:
          $ref: '#/components/schemas/UserSummary'
        allowedActions:
          type: array
          description: 当前用户在当前状态可执行的订单操作；后端在处理请求时仍重新校验
          items:
            $ref: '#/components/schemas/OrderAction'
        events:
          type: array
          items:
            $ref: '#/components/schemas/OrderEvent'
        createdAt:
          $ref: '#/components/schemas/UtcInstant'
        confirmedAt:
          anyOf:
            - $ref: '#/components/schemas/UtcInstant'
            - type: 'null'
        deliveredAt:
          anyOf:
            - $ref: '#/components/schemas/UtcInstant'
            - type: 'null'
        completedAt:
          anyOf:
            - $ref: '#/components/schemas/UtcInstant'
            - type: 'null'
        cancelledAt:
          anyOf:
            - $ref: '#/components/schemas/UtcInstant'
            - type: 'null'
        updatedAt:
          $ref: '#/components/schemas/UtcInstant'

    # -------------------------------------------------------------------------
    # 评价
    # -------------------------------------------------------------------------
...
    AdminOrderSummary:
      type: object
      description: 管理端全站订单条目，同时返回买卖双方；与个人订单的 counterpart 视角分离
      required: [id, orderNo, status, tradeMode, amount, item, buyer, seller, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/DecimalId'
        orderNo:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        tradeMode:
          $ref: '#/components/schemas/TradeMode'
        amount:
          $ref: '#/components/schemas/MoneyAmount'
        item:
          $ref: '#/components/schemas/OrderItemSnapshot'
        buyer:
          $ref: '#/components/schemas/UserSummary'
        seller:
          $ref: '#/components/schemas/UserSummary'
        createdAt:
          $ref: '#/components/schemas/UtcInstant'
        updatedAt:
          $ref: '#/components/schemas/UtcInstant'
...